add all quizzes report

+++ src/main/java/com/letslive/letslearnbackend/dto/QuestionDTO.java
@@ -1,19 +1,11 @@
 package com.letslive.letslearnbackend.dto;
 
-import com.letslive.letslearnbackend.entities.Course;
-import com.letslive.letslearnbackend.entities.User;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Set;
 import java.util.UUID;
 
 @Getter
@@ -25,7 +17,7 @@ public class QuestionDTO {
     private String questionText;
     private String status;
     private String type;
-    private BigDecimal defaultMark;
+    private Double defaultMark;
     private Long usage;
     private String feedbackOfTrue;
     private String feedbackOfFalse;
